Introducing a buffered reader seems to improve the performance of decoding packets substantially. 

The benchmark that this PR introduces is somewhat flawed in that it measures reads from disk instead of from the network, but the reasoning was that reading byte-by-byte through syscalls is most easily replicated this way. 

```
$ benchstat old.txt new.txt
name          old time/op  new time/op  delta
ReadPacket-8   324ms ± 1%    59ms ± 1%  -81.71%  (p=0.002 n=6+6)
```
Benchmark on my machine before:

```
goos: linux
goarch: amd64
pkg: github.com/go-asn1-ber/asn1-ber
BenchmarkReadPacket
BenchmarkReadPacket-8   	       4	 321296674 ns/op
PASS
ok  	github.com/go-asn1-ber/asn1-ber	3.468s
```

and after:
```
goos: linux
goarch: amd64
pkg: github.com/go-asn1-ber/asn1-ber
BenchmarkReadPacket
BenchmarkReadPacket-8   	      19	  59258121 ns/op
PASS
ok  	github.com/go-asn1-ber/asn1-ber	2.349s
```